Refactor test files
Refactor test files
  - merge test files
  - improve test debug messages
